refactor: Update response to align with revised API
