+++ app/src/main/java/com/okaka/challengeonairandroid/model/api/response/HistoryResponse.kt
@@ -1,23 +1,17 @@
 package com.okaka.challengeonairandroid.model.api.response
 
+import com.challengeonair.challengeonairandroid.model.data.entity.Challenge
 import com.google.gson.annotations.SerializedName
 
-data class ChallengeHistoryResponse(
-    @SerializedName("challenge_name") val challengeName: String,
-    @SerializedName("challenge_date") val challengeDate: String,
-    @SerializedName("startTime") val startTime: String,
-    @SerializedName("endTime") val endTime: String,
-    @SerializedName("imageUrl") val imageUrl: String,
-    @SerializedName("participantNum") val participantNum: Int
-)
-
+// 1. 회원 챌린지 History 단건 조회
 data class HistoryResponse(
-    @SerializedName("challenge") val challenge: ChallengeResponse,
+    @SerializedName("challenge") val challenge: Challenge,
     @SerializedName("isSucceeded") val isSucceeded: Boolean,
     @SerializedName("isHost") val isHost: Boolean,
     @SerializedName("point") val point: Int
 )
 
+// 2. 회원 챌린지 History 전체 조회
 data class AllHistoriesResponse(
     @SerializedName("histories") val histories: List<HistoryResponse>
 )
