+++ app/src/main/java/com/okaka/challengeonairandroid/model/api/response/UserProfileResponse.kt
@@ -2,19 +2,20 @@ package com.okaka.challengeonairandroid.model.api.response
 
 import com.google.gson.annotations.SerializedName
 
+// 1. 회원 정보 조회
 data class UserProfileResponse(
-    @SerializedName("user_nick_name") val userNickName: String,
-    @SerializedName("user_body") val userBody: String,
-    @SerializedName("image_url") val imageUrl: String,
+    @SerializedName("userNickName") val userNickName: String,
+    @SerializedName("imageUrl") val imageUrl: String,
     @SerializedName("point") val point: Int
 )
 
+// 2. 회원 정보 수정
 data class UserProfileUpdateRequest(
-    @SerializedName("user_nick_name") val userNickName: String? = null,
-    @SerializedName("user_body") val userBody: String? = null,
-    @SerializedName("image_url") val imageUrl: String? = null
+    @SerializedName("userNickName") val userNickName: String? = null,
+    @SerializedName("imageUrl") val imageUrl: String? = null
 )
 
 data class UserProfileUpdateResponse(
-    @SerializedName("user_id") val userId: String
+    @SerializedName("userNickName") val userNickName: String,
+    @SerializedName("imageUrl") val imageUrl: String
 )
