+++ app/src/main/java/com/okaka/challengeonairandroid/model/data/entity/Category.kt
@@ -1,0 +1,8 @@
+package com.okaka.challengeonairandroid.model.data.entity
+
+enum class Category(val id: Long) {
+    EXERCISE(1L),
+    DEVELOPMENT(2L),
+    HOBBY(3L),
+    STUDY(4L)
+}
