+++ app/src/main/java/com/okaka/challengeonairandroid/view/mypage/MyPageHistoryActivity.kt
@@ -27,10 +27,6 @@ class MyPageHistoryActivity : AppCompatActivity() {
 
         val rvHistory = myPageHistoryBinding.rvHistory
 
-        val dummyHistoryList = listOf(
-            History("하루 30분 모각독", "11:00", "11:30", "08.24", isSucceeded = true, isHost = true),
-            History("방 처음 파본다", "10:30", "11:00", "08.25", isSucceeded = true, isHost = true)
-        )
 
         val btnBack = myPageHistoryBinding.btnBack
         btnBack.setOnClickListener {
